requirements.txt
torch
transformers
datasets
peft
accelerate
bitsandbytes
tqdm
pyyaml
matplotlib
